Move TestOpenBlockSeriesChunkRefsSetsIterator
