+++ pkg/storegateway/batch_series_test.go
@@ -9,15 +9,12 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/go-kit/log"
 	"github.com/oklog/ulid"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb/chunks"
-	"github.com/prometheus/prometheus/tsdb/hashcache"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/thanos-io/objstore"
 	"go.uber.org/atomic"
 
 	"github.com/grafana/mimir/pkg/storegateway/storepb"
@@ -53,166 +50,6 @@ func TestPreloadingBatchSet_Concurrency(t *testing.T) {
 		require.Error(t, preloading.Err())
 	}
 
-}
-
-func TestOpenBlockSeriesChunkRefsSetsIterator(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	t.Cleanup(cancel)
-
-	testCases := map[string]struct {
-		matcher        *labels.Matcher
-		batchSize      int
-		chunksLimit    int
-		seriesLimit    int
-		expectedErr    string
-		expectedSeries []seriesChunkRefsSet
-	}{
-		"chunks limits reached": {
-			matcher:     labels.MustNewMatcher(labels.MatchRegexp, "a", ".+"),
-			batchSize:   100,
-			chunksLimit: 1,
-			seriesLimit: 100,
-			expectedErr: "test limit exceeded",
-		},
-		"series limits reached": {
-			matcher:     labels.MustNewMatcher(labels.MatchRegexp, "a", ".+"),
-			batchSize:   100,
-			chunksLimit: 100,
-			seriesLimit: 1,
-			expectedErr: "test limit exceeded",
-		},
-		"selects all series in a single batch": {
-			matcher:     labels.MustNewMatcher(labels.MatchRegexp, "a", ".+"),
-			batchSize:   100,
-			chunksLimit: 100,
-			seriesLimit: 100,
-			expectedSeries: []seriesChunkRefsSet{
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "1")},
-					{lset: labels.FromStrings("a", "1", "b", "2")},
-					{lset: labels.FromStrings("a", "2", "b", "1")},
-					{lset: labels.FromStrings("a", "2", "b", "2")},
-				}},
-			},
-		},
-		"selects all series in multiple batches": {
-			matcher:     labels.MustNewMatcher(labels.MatchRegexp, "a", ".+"),
-			batchSize:   1,
-			chunksLimit: 100,
-			seriesLimit: 100,
-			expectedSeries: []seriesChunkRefsSet{
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "1")},
-				}},
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "2")},
-				}},
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "2", "b", "1")},
-				}},
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "2", "b", "2")},
-				}},
-			},
-		},
-		"selects some series in single batch": {
-			matcher:     labels.MustNewMatcher(labels.MatchEqual, "a", "1"),
-			batchSize:   100,
-			chunksLimit: 100,
-			seriesLimit: 100,
-			expectedSeries: []seriesChunkRefsSet{
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "1")},
-					{lset: labels.FromStrings("a", "1", "b", "2")},
-				}},
-			},
-		},
-		"selects some series in multiple batches": {
-			matcher:     labels.MustNewMatcher(labels.MatchEqual, "a", "1"),
-			batchSize:   1,
-			chunksLimit: 100,
-			seriesLimit: 100,
-			expectedSeries: []seriesChunkRefsSet{
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "1")},
-				}},
-				{series: []seriesChunkRefs{
-					{lset: labels.FromStrings("a", "1", "b", "2")},
-				}},
-			},
-		},
-	}
-
-	for testName, testCase := range testCases {
-		testName, testCase := testName, testCase
-		t.Run(testName, func(t *testing.T) {
-			t.Parallel()
-
-			suite := prepareStoreWithTestBlocks(t, t.TempDir(), objstore.NewInMemBucket(), false, NewChunksLimiterFactory(0), NewSeriesLimiterFactory(0))
-			var firstBlock *bucketBlock
-			// Find the block with the smallest timestamp in its ULID.
-			// The test setup creates two blocks - each takes 4 different timeseries; each has
-			// a timestamp of time.Now() when its being created.
-			// We want the first created block because we want to assert on the series inside it.
-			// The block created first contains a known set of 4 series.
-			// TODO dimitarvdimitrov change this setup to use prepareTestBlock() and assert on the chunks too
-			for _, b := range suite.store.blocks {
-				if firstBlock == nil {
-					firstBlock = b
-					continue
-				}
-				if b.meta.ULID.Time() < firstBlock.meta.ULID.Time() {
-					firstBlock = b
-				}
-			}
-			suite.cache.SwapWith(noopCache{})
-
-			indexReader := firstBlock.indexReader()
-			defer indexReader.Close()
-
-			iterator, err := openBlockSeriesChunkRefsSetsIterator(
-				ctx,
-				testCase.batchSize,
-				indexReader,
-				firstBlock.meta,
-				[]*labels.Matcher{testCase.matcher},
-				nil,
-				hashcache.NewSeriesHashCache(1024*1024).GetBlockCache(firstBlock.meta.ULID.String()),
-				&limiter{limit: testCase.chunksLimit},
-				&limiter{limit: testCase.seriesLimit},
-				false,
-				firstBlock.meta.MinTime,
-				firstBlock.meta.MaxTime,
-				newSafeQueryStats(),
-				log.NewNopLogger(),
-			)
-			require.NoError(t, err)
-
-			actualSeriesSets := readAllSeriesChunkRefsSet(iterator)
-
-			require.Lenf(t, actualSeriesSets, len(testCase.expectedSeries), "expected %d sets, but got %d", len(testCase.expectedSeries), len(actualSeriesSets))
-			for i, actualSeriesSet := range actualSeriesSets {
-				expectedSeriesSet := testCase.expectedSeries[i]
-				require.Equal(t, expectedSeriesSet.len(), actualSeriesSet.len())
-				for j, actualSeries := range actualSeriesSet.series {
-					expectedSeries := testCase.expectedSeries[i].series[j]
-
-					actualLset := actualSeries.lset
-					expectedLset := expectedSeries.lset
-					assert.Truef(t, labels.Equal(actualLset, expectedLset), "%d, %d: expected labels %s got labels %s", i, j, expectedLset, actualLset)
-
-					// We can't test anything else from the chunk ref because it is generated on the go in each test case
-					assert.Len(t, actualSeries.chunks, 1)
-					assert.Equal(t, firstBlock.meta.ULID, actualSeries.chunks[0].blockID)
-				}
-			}
-			if testCase.expectedErr != "" {
-				assert.ErrorContains(t, iterator.Err(), "test limit exceeded")
-			} else {
-				assert.NoError(t, iterator.Err())
-			}
-		})
-	}
 }
 
 func TestLoadingBatchSet(t *testing.T) {
